refactor(ui): extract event API functions into dedicated events.ts
Move event-related functions (listEvents, getEventStreamUrl) from
messages.ts to their own events.ts file for proper code organization.
Events are a separate concern from messages.
