+++ apps/ui/src/lib/api/index.ts
@@ -1,8 +1,9 @@
-// API exports (M2)
+// API exports
 export * from "./types";
 export * from "./client";
 export * from "./agents";
 export * from "./sessions";
 export * from "./messages";
+export * from "./events";
 export * from "./llm-providers";
 export * from "./auth";
